Use the `:once` VCR recording mode
The big change here is using `record` options of :once rather than :new_episodes.  Now that we are using the RSpec metadata to produce a fixture for each test, it is recommended to use :once which either allows real HTTP connections OR stubs the requests in the cassette (but not both).  Which it does is based on whether or not there is a cassette file or not.  If there is one, it does playback but not recording; if there is not a file, it does recording but not playback.

The development cycle changes a bit here as it is required to be more intentional in creating the cassette.  The benifits include knowing exactly when the real http connections are being made.

`VCR.current_cassette.recording?` will only work (and really only makes sense) if record is set to :once.   As such we can do things like wait for solr to do some indexing when recording a new cassette, but keep the specs under a half second durring normal test runs.