Fixed so errors are properly constructed.
ActiveModelErrors#add was being called with an array containing a string when
it should have just been called with a string.  See
http://api.rubyonrails.org/classes/ActiveModel/Errors.html#method-i-add
for more info.

I have verified that bin/rake still passes with this mod.

I have not added a new spec because this gem is already lacking
coverage and a spec would mostly be a feel-good exercise, because the
functionality that has been changed only makes a difference to a
client that is using this gem.

If this gem already had 100% coverage, I'd add a spec because having
100% coverage is like double-entry accounting; 100% provides
protection from a certain class of errors, but it's much less useful
if done incompletely.